pyproject.toml
[project]
name = "time-tracker"
version = "0.1.0"
description = "Multi-client concurrent time tracker"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=10.0.0",
    "pyobjc-framework-cocoa>=10.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
    "ty>=0.0.25",
]


